+++ contracts/MKNFT4.sol
@@ -3,25 +3,23 @@
 pragma solidity ^0.8.0;
 
 import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
-import "@openzeppelin/contracts/utils/Counters.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 
 contract MKNFT4 is ERC721URIStorage, Ownable {
-    using Counters for Counters.Counter;
-    Counters.Counter private _tokenIds;
+    uint256 private _tokenIds;
 
-    constructor() ERC721("MK Token4 for NFT", "MKNFT4") {}
+    constructor() ERC721("MK Token4 for NFT", "MKNFT4") Ownable(msg.sender) {}
 
     function mintNFT(address recipient, string memory tokenURI)
         public
         onlyOwner
         returns (uint256)
     {
-        require(_tokenIds.current() <= 10);
-        _tokenIds.increment();
+        require(_tokenIds <= 10);
+        _tokenIds++;
 
-        uint256 newItemId = _tokenIds.current();
+        uint256 newItemId = _tokenIds;
         _mint(recipient, newItemId);
         _setTokenURI(newItemId, tokenURI);