+++ scripts/mint-721.js
@@ -5,8 +5,12 @@ const API_URL = process.env.MUMBAI_API_URL;
 const PUBLIC_KEY = process.env.PUBLIC_KEY;
 const PRIVATE_KEY = process.env.PRIVATE_KEY;
 
-const { createAlchemyWeb3 } = require("@alch/alchemy-web3");
-const web3 = createAlchemyWeb3(API_URL);
+const { Alchemy, Network } = require("alchemy-sdk");
+const alchemy = new Alchemy({
+    apiKey: process.env.ALCHEMY_API_KEY,
+    network: Network.MATIC_MUMBAI,
+});
+const web3 = alchemy.core;
 
 console.log("Connected with", API_URL);